fix: 정속주행률 데이터타입 변환으로 인한 에러 해결

+++ src/main/java/com/smooth/driving_analysis_service/reports/basic_summary/service/BasicSummaryServiceImpl.java
@@ -128,7 +128,7 @@ public class BasicSummaryServiceImpl implements BasicSummaryService {
                 .averageDurationSec(toBigDecimal(projection.getAverageDurationSec()))
                 .averageDistanceKm(toBigDecimal(projection.getAverageDistanceKm()))
                 .averageSpeedKmh(toBigDecimal(projection.getAverageSpeedKmh()))
-                .averageCruiseRatio(toBigDecimal(projection.getAverageCruiseRatio()))
+                .averageCruiseRatio(toBigDecimal((double) projection.getAverageCruiseRatio()))
                 .snapshotType(snapshotType)
                 .build();
